+++ ci/cpu/libcudf/build_libcudf.sh
@@ -9,8 +9,8 @@ if [ "$BUILD_LIBCUDF" == '1' -o "$BUILD_CFFI" == '1' ]; then
   fi
 
   if [ "$BUILD_ABI" == "1" ]; then
-    conda build conda/recipes/libcudf -c nvidia/label/cuda${CUDA_REL} -c rapidsai/label/cuda${CUDA_REL} -c numba -c conda-forge -c defaults
+    conda build conda/recipes/libcudf -c nvidia/label/cuda${CUDA_REL} -c rapidsai/label/cuda${CUDA_REL} -c rapidsai-nightly/label/cuda${CUDA_REL} -c numba -c conda-forge -c defaults
   else
-    conda build conda/recipes/libcudf -c nvidia/label/cf201901-cuda${CUDA_REL} -c rapidsai/label/cf201901-cuda${CUDA_REL} -c numba -c conda-forge/label/cf201901 -c defaults
+    conda build conda/recipes/libcudf -c nvidia/label/cf201901-cuda${CUDA_REL} -c rapidsai/label/cf201901-cuda${CUDA_REL} -c rapidsai-nightly/label/cf201901-cuda${CUDA_REL} -c numba -c conda-forge/label/cf201901 -c defaults
   fi
 fi