+++ python/cudf/tests/test_dataframe.py
@@ -1202,18 +1202,22 @@ def test_to_from_arrow_nulls(data_type):
         s1 = pa.array([1, None, 3, None, 5], type=data_type)
     gs1 = gd.Series.from_arrow(s1)
     assert isinstance(gs1, gd.Series)
+    # We have 64B padded buffers for nulls whereas Arrow returns a minimal
+    # number of bytes, so only check the first byte in this case
     np.testing.assert_array_equal(
-        np.array(s1.buffers()[0]),
-        gs1.nullmask.to_array()
+        np.array(s1.buffers()[0])[0],
+        gs1.nullmask.to_array()[0]
     )
     assert pa.Array.equals(s1, gs1.to_arrow())
 
     s2 = pa.array([None, None, None, None, None], type=data_type)
     gs2 = gd.Series.from_arrow(s2)
     assert isinstance(gs2, gd.Series)
+    # We have 64B padded buffers for nulls whereas Arrow returns a minimal
+    # number of bytes, so only check the first byte in this case
     np.testing.assert_array_equal(
-        np.array(s2.buffers()[0]),
-        gs2.nullmask.to_array()
+        np.array(s2.buffers()[0])[0],
+        gs2.nullmask.to_array()[0]
     )
     assert pa.Array.equals(s2, gs2.to_arrow())